Manejo de errores en el backend

+++ spring-boot-backend-apirest/src/main/java/com/example/springbootbackendapirest/controllers/ClienteRestController.java
@@ -1,13 +1,16 @@
 package com.example.springbootbackendapirest.controllers;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.springbootbackendapirest.models.entity.Cliente;
@@ -15,6 +18,10 @@ import com.example.springbootbackendapirest.models.services.IClienteService;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+
+//IMPORTANTE Manejo de errores
 
 @CrossOrigin(origins = {"http://localhost:4200"})
 @RestController 
@@ -31,34 +38,103 @@ public class ClienteRestController {
 
     }
 
+    // Lo que estamos cambiando aca es que se le pueda mandar un mensaje en caso
+    // tal que no encontremos el id de ese cliente en la bd
     @GetMapping("/clientes/{id}")
-    public Cliente show(@PathVariable Long id){
-        return clienteService.findById(id);
+    public ResponseEntity<?> show(@PathVariable Long id){ // el ? indica cualquier tipo de objeto
+        
+        Cliente cliente = null;
+        Map<String, Object> response = new HashMap<>();
+
+        try {
+            cliente = clienteService.findById(id);
+        } catch (DataAccessException e) {
+            response.put("mensaje", "Error al realizar la consulta en la base de datos");
+            response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+            return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        
+        
+        if (cliente ==  null){
+            response.put("mensaje", "El cliente ID: ".concat(id.toString().concat(" no existe en la base de datos")));
+            return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<Cliente>(cliente, HttpStatus.OK);
     }
 
+
     @PostMapping("/clientes")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Cliente create(@RequestBody Cliente cliente){
-        return clienteService.save(cliente);
+    public ResponseEntity<?> create(@RequestBody Cliente cliente){
+
+        Cliente clienteNew = null;
+        Map<String, Object> response = new HashMap<>();
+
+        try {
+            clienteNew = clienteService.save(cliente);
+        } catch (DataAccessException e) {
+            response.put("mensaje", "Error al realizar el insert en la base de datos");
+            response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+            return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        response.put("mensaje", "El cliente ha sido creado con éxito!");
+        response.put("cliente", clienteNew);
+        return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
+
+
     }
 
     @PutMapping("clientes/{id}")
-    @ResponseStatus(HttpStatus.CREATED)
-    public Cliente update(@RequestBody Cliente cliente, @PathVariable Long id){
+    public ResponseEntity<?> update(@RequestBody Cliente cliente, @PathVariable Long id){
+
+        
         Cliente clienteActual = clienteService.findById(id);
+        Cliente clienteUpdated = null;
+        Map<String, Object> response = new HashMap<>();
 
-        clienteActual.setNombre(cliente.getNombre());
-        clienteActual.setApellido(cliente.getApellido());
-        clienteActual.setEmail(cliente.getEmail());
+        if (clienteActual == null){
+            response.put("mensaje", "Error: no se pudo editar, el cliente ID: ".concat(id.toString().concat(" no existe en la base de datos")));
+            return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND); 
+        }
 
-        return clienteService.save(clienteActual);
+        try {
+            clienteActual.setNombre(cliente.getNombre());
+            clienteActual.setApellido(cliente.getApellido());
+            clienteActual.setEmail(cliente.getEmail());
+            clienteUpdated = clienteService.save(clienteActual);
+        } catch (DataAccessException e){
+            response.put("mensaje", "Error al realizar actualizar el cliente en la base de datos");
+            response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+            return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        
+        response.put("mensaje","El cliente ha sido actualizado con éxito!");
+        response.put("cliente", clienteUpdated);
+        return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
     } 
 
 
     @DeleteMapping("/clientes/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void update(@PathVariable Long id) {        
-        clienteService.delete(id);
+    public ResponseEntity<?> update(@PathVariable Long id) {        
+
+        Map<String, Object> response = new HashMap<>();
+
+        Cliente clienteDelete = clienteService.findById(id);
+ 
+        if (clienteDelete == null){
+            response.put("mensaje", "No se pudo eliminar, ya que el cliente no está en la base de datos");
+            return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+        }
+
+        try {
+            clienteService.delete(id);
+        } catch (DataAccessException e){
+            response.put("mensaje", "Error al eliminar el cliente en la base de datos");
+            response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+            return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+        response.put("mensaje","El cliente ha sido eliminado con éxito!");
+        return new ResponseEntity<Map<String, Object>>(response, HttpStatus.OK);
     }
 
 }
